backend/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
asyncpg